Move Badge Renderer to lib
