+++ src/app/api/badge/route.ts
@@ -1,5 +1,5 @@
 import { getUserCount } from "@/lib/tracking";
-import { renderBadge, resolveBadgeStyle } from "./badge-svg";
+import { renderBadge, resolveBadgeStyle } from "../../../lib/svg/badge";
 import { getCacheHeaders } from "@/lib/cache";
 
 export const dynamic = "force-dynamic";
@@ -8,7 +8,12 @@ export async function GET(request: Request) {
   const url = new URL(request.url);
   const style = resolveBadgeStyle(url.searchParams.get("style"));
   const count = await getUserCount();
-  const svg = renderBadge("users", count.toLocaleString("en-US"), "4c8eda", style);
+  const svg = renderBadge(
+    "users",
+    count.toLocaleString("en-US"),
+    "4c8eda",
+    style,
+  );
 
   return new Response(svg, {
     status: 200,
