+++ src/lib/svg/badge/configs/styleConfig.ts
@@ -1,0 +1,81 @@
+import { BadgeStyle, BadgeStyleConfig } from "../types/types";
+
+export const STYLE_CONFIGS: Record<BadgeStyle, BadgeStyleConfig> = {
+  flat: {
+    height: 20,
+    charWidth: 6.6,
+    pad: 10,
+    rx: 3,
+    fontSize: 11,
+    fontWeight: 400,
+    letterSpacing: 0,
+    textY: 14,
+    uppercase: false,
+    transparent: false,
+    gradient: "shields",
+    clipId: "r_flat",
+    gradientId: "s_flat",
+  },
+  "flat-square": {
+    height: 20,
+    charWidth: 6.6,
+    pad: 10,
+    rx: 0,
+    fontSize: 11,
+    fontWeight: 400,
+    letterSpacing: 0,
+    textY: 14,
+    uppercase: false,
+    transparent: false,
+    gradient: "shields",
+    clipId: "r_sq",
+    gradientId: "s_sq",
+  },
+  "for-the-badge": {
+    height: 28,
+    charWidth: 7.6,
+    pad: 16,
+    rx: 0,
+    fontSize: 11,
+    fontWeight: 900,
+    letterSpacing: 0.5,
+    textY: 20,
+    uppercase: true,
+    transparent: false,
+    gradient: "none",
+    clipId: "r_ftb",
+    gradientId: "g_ftb",
+  },
+  plastic: {
+    height: 20,
+    charWidth: 6.6,
+    pad: 10,
+    rx: 3,
+    fontSize: 11,
+    fontWeight: 400,
+    letterSpacing: 0,
+    textY: 14,
+    uppercase: false,
+    transparent: false,
+    gradient: "plastic",
+    clipId: "r_pl",
+    gradientId: "s_pl",
+  },
+  minimal: {
+    height: 20,
+    charWidth: 6.6,
+    pad: 10,
+    rx: 0,
+    fontSize: 11,
+    fontWeight: 400,
+    letterSpacing: 0,
+    textY: 14,
+    uppercase: false,
+    transparent: true,
+    gradient: "none",
+    clipId: "r_min",
+    gradientId: "g_min",
+  },
+};
+
+export const VALID_STYLES = new Set<string>(Object.keys(STYLE_CONFIGS));
